+++ CHANGELOG.md
@@ -16,9 +16,9 @@
 
 ### Master
 
-
 - Requires phone number during bidder registration - sweir27
 - Migrates to React Native 0.57.x - orta/alloy/luc/justin
+- Adds correct margin/padding to last carousel image - kieran
 
 ### 1.9.1