+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Master
 
+- Fixes City not re-rendering after a MP request is successfully retried - ash
 - Fixes Closing Soon and Opening Soon order - kieran
 
 ### 1.9.2