+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 ### Master
 
 - Reduces Emission's City cache TTL to 2 hours - kieran
+- Fixes City not re-rendering after a MP request is successfully retried - ash
+- Fixes Closing Soon and Opening Soon order - kieran
 
 ### 1.9.2