+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Fixes City not re-rendering after a MP request is successfully retried - ash
 - Fixes Closing Soon and Opening Soon order - kieran
+- Fairbooths now only pull 5 shows per page on connection - kieran
 
 ### 1.9.2