+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Master
 
+### 1.9.3
+
 - Reduces Emission's City cache TTL to 2 hours - kieran
 - Fixes City not re-rendering after a MP request is successfully retried - ash
 - Fixes Closing Soon and Opening Soon order - kieran